+++ IOStream/FileInputStream.h
@@ -1,5 +1,5 @@
 /*
- * Copyright 2011 Jason Rush and John Flanagan. All rights reserved.
+ * Copyright 2011-2012 Jason Rush and John Flanagan. All rights reserved.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -22,7 +22,7 @@
     int fd;
 }
 
-- (id)initWithFilename:(NSString*)filename;
-- (NSInteger)seek:(NSUInteger)offset;
+- (id)initWithFilename:(NSString *)filename;
+- (off_t)seek:(off_t)offset;
 
 @end
