+++ KeePassLib/DDXMLDocument+MKPAdditions.m
@@ -1,10 +1,19 @@
-//
-//  DDXMLDocument+MKPAdditions.m
-//  MiniKeePass
-//
-//  Created by Jason Rush on 9/15/12.
-//  Copyright (c) 2012 Self. All rights reserved.
-//
+/*
+ * Copyright 2011-2012 Jason Rush and John Flanagan. All rights reserved.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
 
 #import "DDXMLDocument+MKPAdditions.h"
 
@@ -18,7 +27,7 @@
 	// Therefore, we call it again here just to be safe.
 	xmlKeepBlanksDefault(0);
 
-	xmlDocPtr doc = xmlReadIO(ioread, ioclose, ioctx, NULL, NULL, mask);
+	xmlDocPtr doc = xmlReadIO(ioread, ioclose, ioctx, NULL, NULL, (int)mask);
 	if (doc == NULL) {
 		if (error) {
             *error = [NSError errorWithDomain:@"DDXMLErrorDomain" code:1 userInfo:nil];
